requirements.txt
# openrobot MVP dependencies

# Simulation
mujoco>=3.1.0
mediapy>=1.1.0

# Perception & VLA
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.38.0
opencv-python>=4.8.0
pillow>=10.0.0

# Cognition & Memory
openai>=1.12.0
chromadb>=0.4.0
sentence-transformers>=2.3.0

# Spatial & Geometry
numpy>=1.24.0
scipy>=1.11.0

# Utilities
pyyaml>=6.0
requests>=2.31.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
